+++ test/util/zod.ts
@@ -1,0 +1,5 @@
+import { ZodError } from "zod";
+
+export const zodError = (message: string, prop: string) => {
+  return new ZodError([{ code: "custom", message, path: [prop] }]);
+};